Add collapsible note and remove Sidebar component
A lot of things are going on here. I changed my forms to make sure
they automatically match the height of the note when editing. Also
when a note is collapsed it will automatically exit edit/delete state.
The editing/adding/delete state was carried down from NotesContext to
the note component. The Noteform component was also integrated into
the Note component so that each note can be edited. I may be missing
a couple more, going to do better with atomic commits!