Calibrate react-hook-form to work with ref and handleChange
React-hook-form has it's own functions to track ref and onchange
that are passed to the input/textarea to track things like changes
to the field. I needed to make sure the text area had both the
textAreaRef to auto-adjust height, but also both the onChange and
handleChange functions passed to each field to make sure I can
pull data in the end from onSubmit.